add: validation checks in services and team init at deployment

+++ backend/src/controllers/auth.controller.js
@@ -8,26 +8,35 @@ const { generateToken, verifyToken } = require("../utils/jwt");
 const crypto = require('crypto');
 const { TOKEN_LIFESPAN } = require('../utils/constants');
 const { sendSignupEmail, sendPasswordResetEmail } = require('../service/email.service');
+const settings = require("../../config/settings");
 
 const register = async (req, res) => {
+  const transaction = await sequelize.transaction();
   try {
     const { name, surname, email, password } = req.body;
     const existingUser = await User.findOne({ where: { email } });
     if (existingUser) return res.status(400).json({ error: "Email already exists" });
 
-    const invite = await Invite.findOne({
-      where: { invitedEmail: email }
-    })
-    if(!invite) {
-      return res.status(400).json({ error: "No Invite Found" });
-    }
-
-    const transaction = await sequelize.transaction();
-    await invite.destroy({ transaction });
-
+    const userCount = await User.count();
     const hashedPassword = await bcrypt.hash(password, 10);
-    const newUser = await User.create({ name, surname, email, password: hashedPassword, role: invite.role }, transaction);
-    await transaction.commit();
+
+    let newUser;
+    if(userCount) {
+      const invite = await Invite.findOne({
+        where: { invitedEmail: email }
+      })
+      if(!invite) {
+        throw new Error("No Invite Found");
+      }
+  
+      await invite.destroy({ transaction });
+  
+      newUser = await User.create({ name, surname, email, password: hashedPassword, role: invite.role }, transaction);
+      await transaction.commit();
+    }
+    else {
+      newUser = await User.create({ name, surname, email, password: hashedPassword, role: settings.user.role.admin });
+    }
 
     const token = generateToken({ id: newUser.id, email: newUser.email });
 
@@ -56,7 +65,7 @@ const login = async (req, res) => {
     const token = generateToken({ id: user.id, email: user.email });
     await Token.create({ token, userId: user.id, type: 'auth' });
 
-    res.status(200).json({ user, token });
+    res.status(200).json({ user: {name: user.name, surname: user.surname, email: user.email, role: settings.user.roleName[user.role]}, token });
   } catch (error) {
     console.error("Error logging in user:", error);
     res.status(500).json({ error: "Internal Server Error" });
